refactor(core): drop dead params and empty pipeline barrels (#1083)
Remove unused transform fabrication at the zero baseline, dead training
warning parameters, and pure re-export hops (layout-helpers split,
bind-layer-validate, buildBatch/batchMarkCount forwards) that added no
seam. Public export names stay put.
