+++ packages/core/src/pipeline/bind-layer-extras.ts
@@ -8,7 +8,7 @@ import type { ColumnTable } from "../table.js";
 import { checkField } from "./bind-layer-check-field.js";
 import { colorBinding } from "./bind-layer-color-binding.js";
 import { styleBinding } from "./bind-layer-style-binding.js";
-import { applyColorOnFillGeomWarning } from "./bind-layer-validate.js";
+import { applyColorOnFillGeomWarning } from "./bind-layer-color-warn.js";
 import type { ColorBinding, PipelineWarning, StyleBinding } from "./types.js";
 import { PipelineError } from "./types.js";
