+++ packages/core/src/pipeline/geometry.ts
@@ -3,24 +3,9 @@
  * typed-array Scene batches. Dispatch + coord flip + mark counting live here;
  * per-geom builders live in geometry-marks / geometry-composites.
  */
-import type { GeometryBatch } from "../scene.js";
-
-import type { LayerFrame, PipelineWarning, ResolvedColorScale } from "./types.js";
-import type { Frame } from "./geometry-shared.js";
-import type { ResolvedStyleScales } from "./geometry-style.js";
-import { dispatchGeometryBatch } from "./geometry-dispatch.js";
-
 export type { Frame } from "./geometry-shared.js";
 export { flipBatchInPlace } from "./geometry-flip.js";
-export { batchMarkCount } from "./geometry-mark-count.js";
-
-export function buildBatch(
-  frame: LayerFrame,
-  fx: Frame,
-  color: ResolvedColorScale | null,
-  fill: ResolvedColorScale | null,
-  styles: ResolvedStyleScales,
-  warnings: PipelineWarning[],
-): GeometryBatch[] {
-  return dispatchGeometryBatch(frame, fx, color, fill, styles, warnings);
-}
+/** Public alias for {@link renderPrimitiveCount} (package export name). */
+export { renderPrimitiveCount as batchMarkCount } from "../candidate-geometry.js";
+/** Scene-batch dispatch entry used by assemble-geometry-batches. */
+export { dispatchGeometryBatch as buildBatch } from "./geometry-dispatch.js";
